Add deprecated versions of generateSpend() function
Add deprecated versions of generateSpend() function as a temporary solution until multiple node identities are implemented
