+++ node/src/test/kotlin/net/corda/node/services/vault/VaultWithCashTest.kt
@@ -15,7 +15,6 @@ import net.corda.finance.*
 import net.corda.finance.contracts.asset.Cash
 import net.corda.finance.contracts.asset.DUMMY_CASH_ISSUER
 import net.corda.finance.contracts.asset.DUMMY_CASH_ISSUER_KEY
-import net.corda.finance.contracts.generateSpend
 import net.corda.finance.contracts.getCashBalance
 import net.corda.finance.schemas.CashSchemaV1
 import net.corda.node.utilities.CordaPersistence
@@ -98,7 +97,7 @@ class VaultWithCashTest : TestDependencyInjectionBase() {
             database.transaction {
                 // A tx that spends our money.
                 val spendTXBuilder = TransactionBuilder(DUMMY_NOTARY)
-                generateSpend(services, spendTXBuilder, 80.DOLLARS, BOB)
+                Cash.generateSpend(services, spendTXBuilder, 80.DOLLARS, BOB)
                 val spendPTX = services.signInitialTransaction(spendTXBuilder, freshKey)
                 notaryServices.addSignature(spendPTX)
             }
@@ -150,7 +149,7 @@ class VaultWithCashTest : TestDependencyInjectionBase() {
             database.transaction {
                 try {
                     val txn1Builder = TransactionBuilder(DUMMY_NOTARY)
-                    generateSpend(services, txn1Builder, 60.DOLLARS, BOB)
+                    Cash.generateSpend(services, txn1Builder, 60.DOLLARS, BOB)
                     val ptxn1 = notaryServices.signInitialTransaction(txn1Builder)
                     val txn1 = services.addSignature(ptxn1, freshKey)
                     println("txn1: ${txn1.id} spent ${((txn1.tx.outputs[0].data) as Cash.State).amount}")
@@ -186,7 +185,7 @@ class VaultWithCashTest : TestDependencyInjectionBase() {
             database.transaction {
                 try {
                     val txn2Builder = TransactionBuilder(DUMMY_NOTARY)
-                    generateSpend(services, txn2Builder, 80.DOLLARS, BOB)
+                    Cash.generateSpend(services, txn2Builder, 80.DOLLARS, BOB)
                     val ptxn2 = notaryServices.signInitialTransaction(txn2Builder)
                     val txn2 = services.addSignature(ptxn2, freshKey)
                     println("txn2: ${txn2.id} spent ${((txn2.tx.outputs[0].data) as Cash.State).amount}")
@@ -309,7 +308,7 @@ class VaultWithCashTest : TestDependencyInjectionBase() {
         database.transaction {
             // A tx that spends our money.
             val spendTXBuilder = TransactionBuilder(DUMMY_NOTARY)
-            generateSpend(services, spendTXBuilder, 80.DOLLARS, BOB)
+            Cash.generateSpend(services, spendTXBuilder, 80.DOLLARS, BOB)
             val spendPTX = notaryServices.signInitialTransaction(spendTXBuilder)
             val spendTX = services.addSignature(spendPTX, freshKey)
             services.recordTransactions(spendTX)
